+++ src/doomstat.c
@@ -207,7 +207,7 @@ int show_save_messages; // CFG-only
 int show_ssg; // CFG-only
 int hud_stats_format;
 int hud_stats_format_map;
-int hud_stats_icons;
+int hud_allow_icons;
 int hud_highlight_weapon;
 int alt_arms;
 int hud_time_teleport;