+++ src/doomstat.h
@@ -592,7 +592,7 @@ enum {
 };
 extern int hud_stats_format;
 extern int hud_stats_format_map;
-extern int hud_stats_icons;
+extern int hud_allow_icons;
 
 extern int hud_highlight_weapon;
 extern int alt_arms;