+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## New Features
 
-None.
+- **Support for powerup-timer icons**
+  - Replaced `hud_stats_icons` with `hud_allow_icons` [1]
 
 ## Changes
 
@@ -9,3 +10,5 @@ None.
 ## Bug Fixes
 
 - **Crash when loading status bars taller than 32px**
+
+**[1].** This will affect existing config files.